// src/types/WeatherApiHour.d.ts
import WeatherApiCondition from './WeatherApiCondition';

export default interface WeatherApiHour {
	time_epoch: number;
	time: string;
	temp_c: number;
	temp_f: number;
	wind_mph: number;
	wind_kph: number;
	wind_degree: number;
	wind_dir: string;
	pressure_mb: number;
	pressure_in: number;
	precip_mm: number;
	precip_in: number;
	snow_cm: number;
	humidity: number;
	cloud: number;
	feelslike_c: number;
	feelslike_f: number;
	windchill_c: number;
	windchill_f: number;
	heatindex_c: number;
	heatindex_f: number;
	dewpoint_c: number;
	dewpoint_f: number;
	will_it_rain: number;
	will_it_snow: number;
	is_day: number;
	vis_km: number;
	vis_miles: number;
	chance_of_rain: number;
	chance_of_snow: number;
	gust_mph: number;
	gust_kph: number;
	uv: number;
	short_rad: number;
	diff_rad: number;
	condition: WeatherApiCondition;
}